Use most recent tag for submodule updates
The overviewer isn't working with the latest minecraft versions. Upon inspection, it seems like this should be fixed with the latest tagged version of the code base - but there still hasn't been an official release. I don't know if they're just not doing releases now, or what's happening. There are hundreds of issues and dozens of PR's in that repo now, so it's looking like the project might be abandoned.